updated flaky test for callback test in-app
